🧪 test(web): consolidate integration test fixtures
