+++ apps/web/src/app/tests/space/space.test.ts
@@ -1,157 +1,13 @@
-import { expect, test, type BrowserContext, type Page } from '@playwright/test'
+import { expect, test } from '../playwright'
 
-import { AccountsFaker } from '../../../../../../packages/core/src/auth/domain/entities/fakers'
-import {
-  PlanetsFaker,
-  StarsFaker,
-} from '../../../../../../packages/core/src/space/domain/entities/fakers'
-import { UsersFaker } from '../../../../../../packages/core/src/profile/domain/entities/fakers'
-import { ServerMock } from '../shared/mocks/ServerMock'
-import type { ServerMockRoute } from '../shared/types/ServerMockRoute'
-
-const USER_ID = '00000000-0000-4000-8000-000000000301'
-const PLANET_ID = '00000000-0000-4000-8000-000000000302'
-const FIRST_STAR_ID = '00000000-0000-4000-8000-000000000303'
-const NEW_STAR_ID = '00000000-0000-4000-8000-000000000304'
-const LAST_UNLOCKED_STAR_ID = '00000000-0000-4000-8000-000000000305'
-const LOCKED_STAR_ID = '00000000-0000-4000-8000-000000000306'
-
-const STAR_NAMES = {
-  first: 'Primeira estrela',
-  new: 'Estrela recém-desbloqueada',
-  lastUnlocked: 'Última estrela desbloqueada',
-  locked: 'Estrela bloqueada',
-}
-
-function createSpaceFakes(overrides: { hasCompletedSpace?: boolean } = {}) {
-  const account = AccountsFaker.fakeDto({
-    id: USER_ID,
-    name: 'Explorador do espaço',
-    email: 'explorador-do-espaco@stardust.dev',
-    isAuthenticated: true,
-  })
-  const user = UsersFaker.fakeDto({
-    id: USER_ID,
-    name: account.name,
-    email: account.email,
-    slug: 'explorador-do-espaco',
-    unlockedStarsIds: [FIRST_STAR_ID, NEW_STAR_ID, LAST_UNLOCKED_STAR_ID],
-    recentlyUnlockedStarsIds: [NEW_STAR_ID],
-    hasCompletedSpace: overrides.hasCompletedSpace ?? false,
-    lastWeekRankingPosition: null,
-  })
-  const firstStar = StarsFaker.fakeDto({
-    id: FIRST_STAR_ID,
-    name: STAR_NAMES.first,
-    number: 1,
-    slug: 'primeira-estrela',
-    isChallenge: false,
-  })
-  const newStar = StarsFaker.fakeDto({
-    id: NEW_STAR_ID,
-    name: STAR_NAMES.new,
-    number: 2,
-    slug: 'estrela-recem-desbloqueada',
-    isChallenge: false,
-  })
-  const lastUnlockedStar = StarsFaker.fakeDto({
-    id: LAST_UNLOCKED_STAR_ID,
-    name: STAR_NAMES.lastUnlocked,
-    number: 3,
-    slug: 'ultima-estrela-desbloqueada',
-    isChallenge: false,
-  })
-  const lockedStar = StarsFaker.fakeDto({
-    id: LOCKED_STAR_ID,
-    name: STAR_NAMES.locked,
-    number: 4,
-    slug: 'estrela-bloqueada',
-    isChallenge: false,
-  })
-  const planet = PlanetsFaker.fakeDto({
-    id: PLANET_ID,
-    name: 'Planeta de testes',
-    icon: 'planet-icon.png',
-    image: 'planet-image.png',
-    stars: [firstStar, newStar, lastUnlockedStar, lockedStar],
-  })
-
-  return { account, user, planet, lastUnlockedStar }
-}
-
-function createAuthenticatedRoutes(
-  fixtures: ReturnType<typeof createSpaceFakes>,
-): ServerMockRoute[] {
-  return [
-    { method: 'GET', path: '/auth/account', status: 200, body: fixtures.account },
-    {
-      method: 'GET',
-      path: `/profile/users/id/${USER_ID}`,
-      status: 200,
-      body: fixtures.user,
-    },
-    { method: 'GET', path: '/profile/achievements', status: 200, body: [] },
-    {
-      method: 'GET',
-      path: '/reporting/feedback/mine/unread-count',
-      status: 200,
-      body: { count: 0 },
-    },
-    {
-      method: 'POST',
-      path: `/profile/achievements/${USER_ID}/observe`,
-      status: 200,
-      body: [],
-    },
-    { method: 'GET', path: '/space/planets', status: 200, body: [fixtures.planet] },
-    { method: 'GET', path: '/shop/insignias', status: 200, body: [] },
-    {
-      method: 'GET',
-      path: '/shop/rockets',
-      status: 200,
-      body: { items: [], totalItemsCount: 0 },
-    },
-    {
-      method: 'GET',
-      path: '/shop/avatars',
-      status: 200,
-      body: { items: [], totalItemsCount: 0 },
-    },
-  ]
-}
-
-async function registerSpaceScenario(
-  page: Page,
-  context: BrowserContext,
-  overrides: { hasCompletedSpace?: boolean } = {},
-) {
-  const fixtures = createSpaceFakes(overrides)
-
-  await context.clearCookies()
-  await context.addCookies([
-    {
-      name: '@stardust:access-token',
-      value: 'space-page-test-token',
-      domain: '127.0.0.1',
-      path: '/',
-    },
-  ])
-  await ServerMock(page).reset()
-  await ServerMock(page).register(createAuthenticatedRoutes(fixtures))
-
-  return fixtures
-}
+import { STAR_NAMES } from '../fixtures/SpaceFixture'
 
 test.describe('/space', () => {
-  test.afterEach(async ({ page }) => {
-    await ServerMock(page).reset()
-  })
-
   test('loads the journey and renders the progress states for each star', async ({
     page,
-    context,
+    space,
   }) => {
-    const fixtures = await registerSpaceScenario(page, context)
+    const fixtures = await space.register()
 
     await page.goto('/space')
 
@@ -161,8 +17,8 @@ test.describe('/space', () => {
     await expect(page.getByRole('button', { name: STAR_NAMES.locked })).toBeDisabled()
   })
 
-  test('does not navigate when a locked star is selected', async ({ page, context }) => {
-    await registerSpaceScenario(page, context)
+  test('does not navigate when a locked star is selected', async ({ page, space }) => {
+    await space.register()
     await page.goto('/space')
 
     const lockedStarButton = page.getByRole('button', { name: STAR_NAMES.locked })
@@ -175,13 +31,14 @@ test.describe('/space', () => {
 
   test('falls back to the lesson when an unlocked star has no challenge', async ({
     page,
-    context,
+    space,
+    serverApp,
   }) => {
-    const fixtures = await registerSpaceScenario(page, context)
+    const fixtures = await space.register()
     const star = fixtures.lastUnlockedStar
 
-    await ServerMock(page).register([
-      ...createAuthenticatedRoutes(fixtures),
+    await serverApp.register([
+      ...space.authenticatedRoutes(fixtures),
       {
         method: 'GET',
         path: `/challenging/challenges/star/${star.id}`,
@@ -225,9 +82,9 @@ test.describe('/space', () => {
 
   test('shows the ending link after the user completes the space', async ({
     page,
-    context,
+    space,
   }) => {
-    await registerSpaceScenario(page, context, { hasCompletedSpace: true })
+    await space.register({ hasCompletedSpace: true })
     await page.goto('/space')
 
     await expect(page.getByRole('link', { name: 'Agradecimentos.' })).toHaveAttribute(
@@ -239,8 +96,10 @@ test.describe('/space', () => {
   test('navigates to space from the protected shop and redirects guests to sign-in', async ({
     page,
     context,
+    space,
+    serverApp,
   }) => {
-    const fixtures = await registerSpaceScenario(page, context)
+    const fixtures = await space.register()
 
     await page.goto('/shop')
     const learnLink = page.getByRole('link', { name: 'Aprender' }).first()
@@ -250,7 +109,7 @@ test.describe('/space', () => {
     await expect(page.getByText(fixtures.planet.name, { exact: true })).toBeVisible()
 
     await context.clearCookies()
-    await ServerMock(page).register([
+    await serverApp.register([
       {
         method: 'GET',
         path: '/auth/account',
